+++ IPTVPlayer/components/playerselector.py
@@ -83,8 +83,8 @@ class PlayerSelectorWidget(Screen):
         markerHeight = 45 + coverHeight
 
         # position of first marker
-        offsetMarkerX = offsetCoverX - (markerWidth - coverWidth) / 2
-        offsetMarkerY = offsetCoverY - (markerHeight - coverHeight) / 2
+        offsetMarkerX = int(offsetCoverX - (markerWidth - coverWidth) / 2)
+        offsetMarkerY = int(offsetCoverY - (markerHeight - coverHeight) / 2)
 
         # how to calculate position of image with indexes indxX, indxY:
         #posX = offsetCoverX + (coverWidth + disWidth) * indxX
@@ -128,7 +128,7 @@ class PlayerSelectorWidget(Screen):
 
         # pagination
         self.pageItemSize = 16
-        self.pageItemStartX = (offsetCoverX + tmpX * numOfCol + offsetCoverX - disWidth - self.numOfPages * self.pageItemSize) / 2
+        self.pageItemStartX = int((offsetCoverX + tmpX * numOfCol + offsetCoverX - disWidth - self.numOfPages * self.pageItemSize) / 2)
         if screenwidth and screenwidth == 1920:
             self.pageItemStartY = 60
         else:
@@ -261,7 +261,7 @@ class PlayerSelectorWidget(Screen):
             self.currLine = (self.numOfLines - 1)
 
         # calculate new page number
-        newPage = self.currLine / self.numOfRow
+        newPage = int(self.currLine / self.numOfRow)
         if newPage != self.currPage:
             self.currPage = newPage
             self.updateIcons()
@@ -322,12 +322,12 @@ class PlayerSelectorWidget(Screen):
             self.lastSelection = self.numOfItems - 1
 
         # numbers of lines
-        self.numOfLines = self.numOfItems / self.numOfCol
+        self.numOfLines = int(self.numOfItems / self.numOfCol)
         if self.numOfItems % self.numOfCol > 0:
             self.numOfLines += 1
 
         # numbers of pages
-        self.numOfPages = self.numOfLines / self.numOfRow
+        self.numOfPages = int(self.numOfLines / self.numOfRow)
         if self.numOfLines % self.numOfRow > 0:
             self.numOfPages += 1
 
@@ -338,7 +338,7 @@ class PlayerSelectorWidget(Screen):
         self.dispY = 0
 
     def updateIconsList(self, rangeList):
-        idx = self.currPage * (self.numOfCol * self.numOfRow)
+        idx = int(self.currPage * (self.numOfCol * self.numOfRow))
         for y in range(1, self.numOfRow + 1):
             for x in range(1, self.numOfCol + 1):
                 if idx >= rangeList[0] and idx <= rangeList[1]:
@@ -348,7 +348,7 @@ class PlayerSelectorWidget(Screen):
                 idx += 1
 
     def updateIcons(self):
-        idx = self.currPage * (self.numOfCol * self.numOfRow)
+        idx = int(self.currPage * (self.numOfCol * self.numOfRow))
         for y in range(1, self.numOfRow + 1):
             for x in range(1, self.numOfCol + 1):
                 strIndex = "cover_%s%s" % (x, y)
@@ -362,14 +362,14 @@ class PlayerSelectorWidget(Screen):
                     self[strIndex].hide()
         x = self.pageItemStartX + self.currPage * self.pageItemSize
         y = self.pageItemStartY
-        self["page_marker"].instance.move(ePoint(x, y))
+        self["page_marker"].instance.move(ePoint(int(x), y))
 
     def setIdx(self, selIdx):
         if selIdx > self.numOfItems:
             selIdx = self.numOfItems
 
         self.dispX = selIdx % self.numOfCol
-        self.currLine = selIdx / self.numOfCol
+        self.currLine = int(selIdx / self.numOfCol)
 
         self.calcMarkerPosX()
         self.calcMarkerPosY()
@@ -405,7 +405,7 @@ class PlayerSelectorWidget(Screen):
         return
 
     def moveMarker(self, prev_idx=0):
-        new_idx = self.currLine * self.numOfCol + self.dispX
+        new_idx = int(self.currLine * self.numOfCol + self.dispX)
 
         if self.reorderingItemSelected:
             if prev_idx != new_idx:
@@ -423,8 +423,8 @@ class PlayerSelectorWidget(Screen):
         imgPosY = self.offsetCoverY + (self.coverHeight + self.disHeight) * self.dispY
 
         # calculate postion of marker for current image
-        x = imgPosX - (self.markerWidth - self.coverWidth) / 2
-        y = imgPosY - (self.markerHeight - self.coverHeight) / 2
+        x = int(imgPosX - (self.markerWidth - self.coverWidth) / 2)
+        y = int(imgPosY - (self.markerHeight - self.coverHeight) / 2)
 
         #x =  30 + self.dispX * 180
         #y = 130 + self.dispY * 125
@@ -453,7 +453,7 @@ class PlayerSelectorWidget(Screen):
                 self.reorderingItemSelected = True
             return
 
-        idx = self.currLine * self.numOfCol + self.dispX
+        idx = int(self.currLine * self.numOfCol + self.dispX)
         PlayerSelectorWidget.LAST_SELECTION[self.groupName] = idx
 
         if idx < self.numOfItems:
