+++ IPTVPlayer/iptvdm/iptvdmapi.py
@@ -395,9 +395,9 @@ class IPTVDMApi():
 
         # calculate downloadedProcent
         if downloadItem.fileSize > 0 and downloadItem.downloadedSize > 0:
-            downloadItem.downloadedProcent = (100 * downloadItem.downloadedSize) / downloadItem.fileSize
+            downloadItem.downloadedProcent = int((100 * downloadItem.downloadedSize) / downloadItem.fileSize)
         elif downloadItem.totalFileDuration > 0 and downloadItem.downloadedFileDuration > 0:
-            downloadItem.downloadedProcent = (100 * downloadItem.downloadedFileDuration) / downloadItem.totalFileDuration
+            downloadItem.downloadedProcent = int((100 * downloadItem.downloadedFileDuration) / downloadItem.totalFileDuration)
         return True
 
     def updateDownloadItemsStatus(self):
